+++ packages/inertia/src/inertia/console/__init__.py
@@ -1,5 +1,0 @@
-"""Inertia console commands."""
-
-from inertia.console.start_ssr import InertiaStartSsrCommand
-
-__all__ = ["InertiaStartSsrCommand"]